kyocera: parade: use int for integer sysfs show values
These four sysfs show helpers pass the local zero-valued result to
sprintf with %d.  Use int rather than size_t so the argument type agrees
with the existing format on both 32-bit and 64-bit kernels.

+++ drivers/input/touchscreen/mediatek/kyocera/parade/ktp_drv_cyttsp5.c
@@ -292,7 +292,7 @@ static ssize_t ktp_drv_name_show(struct device *dev,
 static ssize_t ktp_drv_event_ctrl_show(struct device *dev,
 		struct device_attribute *attr, char *buf)
 {
-	size_t ret = 0;
+	int ret = 0;
 
 	pr_err("%s: START\n", __func__);
 	pr_err("%s: END\n", __func__);
@@ -312,7 +312,7 @@ static ssize_t ktp_drv_event_ctrl_store(struct device *dev,
 static ssize_t ktp_drv_glove_mode_show(struct device *dev,
 		struct device_attribute *attr, char *buf)
 {
-	size_t ret = 0;
+	int ret = 0;
 	pr_err("%s: START\n", __func__);
 	pr_err("%s: END\n", __func__);
 
@@ -331,7 +331,7 @@ static ssize_t ktp_drv_glove_mode_store(struct device *dev,
 static ssize_t ktp_drv_touch_mode_show(struct device *dev,
 		struct device_attribute *attr, char *buf)
 {
-	size_t ret = 0;
+	int ret = 0;
 	pr_err("%s: START\n", __func__);
 	pr_err("%s: END\n", __func__);
 
@@ -350,7 +350,7 @@ static ssize_t ktp_drv_touch_mode_store(struct device *dev,
 static ssize_t ktp_drv_easywake_mode_show(struct device *dev,
 		struct device_attribute *attr, char *buf)
 {
-	size_t ret = 0;
+	int ret = 0;
 	pr_err("%s: START\n", __func__);
 	pr_err("%s: END\n", __func__);
